feat: populate authoritative current-week summary in reading dashboard
GetDashboardAsync now returns a non-null ReadingWeekSummaryDto for the
current Europe/Stockholm ISO week, evaluated by the same weekly-review
query and progression policy used by preview/commit so the dashboard can
never diverge from committed weekly evidence.

- currentWeek: weekKey from the local programme date via ISOWeek,
  completedSessions (Completed + AwaitingFeedback), qualifyingSessions
  (policy qualifying attempts summed across modes), volumeMinutes
  (all-mode effective floor minutes), reviewCommitted from the persisted
  weekly review, completionThreshold = policy increase completion rate
- empty current week is a non-null zero summary once initialized;
  not-initialized behaviour unchanged
- ReadingWeekSummaryDto.CompletionThreshold int -> double (0.8), matching
  the frontend contract and the policy's completion-rate threshold
- service tests: empty week, mixed modes/statuses/constraints (incl.
  no-guilt recovery/constrained volume-only), Stockholm Sunday/Monday
  boundary across the ISO week, committed-review flag; HTTP dashboard
  JSON contract test asserts the non-null zero summary shape

+++ Nostos.Shared/Dtos/ReadingTrainingDtos.cs
@@ -214,12 +214,18 @@ public record ReadingSessionDto(
     DateTime? CompletedAt
 );
 
+// Authoritative current-week summary for the dashboard, derived by the same
+// progression policy that powers the weekly review (Completed +
+// AwaitingFeedback evidence, floor minutes, qualifying semantics). An empty
+// current week is still a non-null zero summary once the programme is
+// initialized. CompletionThreshold is the policy's completion-rate threshold
+// (0.8) a mode must clear before any increase is permitted.
 public record ReadingWeekSummaryDto(
     string WeekKey,
     int CompletedSessions,
     int QualifyingSessions,
     int VolumeMinutes,
-    int CompletionThreshold,
+    double CompletionThreshold,
     bool ReviewCommitted
 );
